FIX #53: move things out of xypath core, mostly
